feat: Optimization of generated `ToString`/`To...Key` methods when using Constant or Guid key parts
When using either Constant or specifically Guid key parts the generated `ToString`/`ToPartitionKeyString`/`ToSortKeyString` implementations can now utilize `string.Create` to create the string in place.

This removes the potential for the string to need to be copied multiple times when it was created via the interpolation.

Guids are only supported when their formatted representation is known to be a fixed length.

This can be expanded on in the future to enable support for other types.