feat: Enable fast path Format/Parse generation for Enums
Adds a capability into the generator to optimze the usage of Enums in Format/Parse code generation.

For each enum we generate a `{EnumName}Helper` file scoped class, this holds methods with optimal implementations so that they can be used in the fast path for Formatting. Additionally, with this change we source generate an implementation of TryParse that's more optimal than the default dotnet impl.

This optimzation applies to all existing `CompositeKey`'s provided they meet the criteria for the fast path generation.