fix: Source Generator now correctly targets all .NET 8 SDK Versions
Previously the source generator was using Roslyn API v4.10 which only worked in dotnet SDK 8.0.3xx, the intention was that it should work in all 8.0.xxx SDK versions.

To resolve this the version of the API has been rolled back to v4.8, this should have no effect if your using >8.0.3xx already but it will allow older SDK versions to work correctly.