feat: Fast path `Format` generation for string properties
Adds a capability into the generator to optimize the usage of Strings in `Format` code generation.

Now instead of falling back to using a string interpolation we check the length of the string ahead of time and generate a more optimal formatting method.

This optimization applies to all existing `CompositeKey`'s provided they meet the updated criteria for the fast path generation.